# openapi.json
{
  "components": {
    "schemas": {
      "AgentRunStageMetadata": {
        "properties": {
          "emitted_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emitted At"
          },
          "lane_index": {
            "default": 0,
            "title": "Lane Index",
            "type": "integer"
          },
          "lane_total": {
            "default": 0,
            "title": "Lane Total",
            "type": "integer"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "sub_question": {
            "default": "",
            "title": "Sub Question",
            "type": "string"
          }
        },
        "required": [
          "stage",
          "status"
        ],
        "title": "AgentRunStageMetadata",
        "type": "object"
      },
      "CitationSourceRow": {
        "properties": {
          "citation_index": {
            "minimum": 1.0,
            "title": "Citation Index",
            "type": "integer"
          },
          "content": {
            "default": "",
            "title": "Content",
            "type": "string"
          },
          "document_id": {
            "default": "",
            "title": "Document Id",
            "type": "string"
          },
          "rank": {
            "minimum": 1.0,
            "title": "Rank",
            "type": "integer"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "source": {
            "default": "",
            "title": "Source",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "citation_index",
          "rank"
        ],
        "title": "CitationSourceRow",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RuntimeAgentRunAsyncCancelResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "status": {
            "const": "success",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "message"
        ],
        "title": "RuntimeAgentRunAsyncCancelResponse",
        "type": "object"
      },
      "RuntimeAgentRunAsyncStartResponse": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "run_id",
          "status"
        ],
        "title": "RuntimeAgentRunAsyncStartResponse",
        "type": "object"
      },
      "RuntimeAgentRunAsyncStatusResponse": {
        "properties": {
          "cancel_requested": {
            "default": false,
            "title": "Cancel Requested",
            "type": "boolean"
          },
          "checkpoint_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Id"
          },
          "decomposition_sub_questions": {
            "items": {
              "type": "string"
            },
            "title": "Decomposition Sub Questions",
            "type": "array"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "interrupt_payload": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Interrupt Payload"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "message": {
            "default": "",
            "title": "Message",
            "type": "string"
          },
          "output": {
            "default": "",
            "title": "Output",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeAgentRunResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "default": "",
            "title": "Run Id",
            "type": "string"
          },
          "stage": {
            "default": "",
            "title": "Stage",
            "type": "string"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/AgentRunStageMetadata"
            },
            "title": "Stages",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "sub_answers": {
            "items": {
              "$ref": "#/components/schemas/SubQuestionAnswer"
            },
            "title": "Sub Answers",
            "type": "array"
          },
          "sub_qa": {
            "items": {
              "$ref": "#/components/schemas/SubQuestionAnswer"
            },
            "title": "Sub Qa",
            "type": "array"
          },
          "sub_question_artifacts": {
            "items": {
              "$ref": "#/components/schemas/SubQuestionArtifacts"
            },
            "title": "Sub Question Artifacts",
            "type": "array"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "RuntimeAgentRunAsyncStatusResponse",
        "type": "object"
      },
      "RuntimeAgentRunControls": {
        "properties": {
          "hitl": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeHitlControl"
              },
              {
                "type": "null"
              }
            ]
          },
          "query_expansion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeQueryExpansionControl"
              },
              {
                "type": "null"
              }
            ]
          },
          "rerank": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeRerankControl"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "RuntimeAgentRunControls",
        "type": "object"
      },
      "RuntimeAgentRunRequest": {
        "properties": {
          "controls": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeAgentRunControls"
              },
              {
                "type": "null"
              }
            ]
          },
          "custom_prompts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeCustomPrompts"
              },
              {
                "type": "null"
              }
            ]
          },
          "query": {
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "runtime_config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeAgentRunRuntimeConfig"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "query"
        ],
        "title": "RuntimeAgentRunRequest",
        "type": "object"
      },
      "RuntimeAgentRunResponse": {
        "properties": {
          "final_citations": {
            "items": {
              "$ref": "#/components/schemas/CitationSourceRow"
            },
            "title": "Final Citations",
            "type": "array"
          },
          "main_question": {
            "default": "",
            "title": "Main Question",
            "type": "string"
          },
          "output": {
            "title": "Output",
            "type": "string"
          },
          "sub_answers": {
            "items": {
              "$ref": "#/components/schemas/SubQuestionAnswer"
            },
            "title": "Sub Answers",
            "type": "array"
          },
          "sub_qa": {
            "items": {
              "$ref": "#/components/schemas/SubQuestionAnswer"
            },
            "title": "Sub Qa",
            "type": "array"
          }
        },
        "required": [
          "output"
        ],
        "title": "RuntimeAgentRunResponse",
        "type": "object"
      },
      "RuntimeAgentRunResumeRequest": {
        "properties": {
          "resume": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "object"
              },
              {
                "$ref": "#/components/schemas/RuntimeSubquestionResumeEnvelope"
              }
            ],
            "default": true,
            "title": "Resume"
          }
        },
        "title": "RuntimeAgentRunResumeRequest",
        "type": "object"
      },
      "RuntimeAgentRunRuntimeConfig": {
        "properties": {
          "query_expansion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeQueryExpansionControl"
              },
              {
                "type": "null"
              }
            ]
          },
          "rerank": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeRerankControl"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "RuntimeAgentRunRuntimeConfig",
        "type": "object"
      },
      "RuntimeCustomPrompts": {
        "properties": {
          "subanswer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subanswer"
          },
          "synthesis": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Synthesis"
          }
        },
        "title": "RuntimeCustomPrompts",
        "type": "object"
      },
      "RuntimeHitlControl": {
        "properties": {
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          },
          "subquestions": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeSubquestionHitlControl"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "RuntimeHitlControl",
        "type": "object"
      },
      "RuntimeQueryExpansionControl": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "title": "RuntimeQueryExpansionControl",
        "type": "object"
      },
      "RuntimeRerankControl": {
        "properties": {
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "title": "RuntimeRerankControl",
        "type": "object"
      },
      "RuntimeSubquestionDecision": {
        "properties": {
          "action": {
            "enum": [
              "approve",
              "edit",
              "deny",
              "skip"
            ],
            "title": "Action",
            "type": "string"
          },
          "edited_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Edited Text"
          },
          "subquestion_id": {
            "minLength": 1,
            "title": "Subquestion Id",
            "type": "string"
          }
        },
        "required": [
          "subquestion_id",
          "action"
        ],
        "title": "RuntimeSubquestionDecision",
        "type": "object"
      },
      "RuntimeSubquestionHitlControl": {
        "properties": {
          "enabled": {
            "default": false,
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "title": "RuntimeSubquestionHitlControl",
        "type": "object"
      },
      "RuntimeSubquestionResumeEnvelope": {
        "properties": {
          "checkpoint_id": {
            "minLength": 1,
            "title": "Checkpoint Id",
            "type": "string"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/RuntimeSubquestionDecision"
            },
            "minItems": 1,
            "title": "Decisions",
            "type": "array"
          }
        },
        "required": [
          "checkpoint_id",
          "decisions"
        ],
        "title": "RuntimeSubquestionResumeEnvelope",
        "type": "object"
      },
      "SubQuestionAnswer": {
        "properties": {
          "answerable": {
            "default": false,
            "title": "Answerable",
            "type": "boolean"
          },
          "expanded_query": {
            "default": "",
            "title": "Expanded Query",
            "type": "string"
          },
          "sub_agent_response": {
            "default": "",
            "title": "Sub Agent Response",
            "type": "string"
          },
          "sub_answer": {
            "title": "Sub Answer",
            "type": "string"
          },
          "sub_question": {
            "title": "Sub Question",
            "type": "string"
          },
          "tool_call_input": {
            "default": "",
            "title": "Tool Call Input",
            "type": "string"
          },
          "verification_reason": {
            "default": "",
            "title": "Verification Reason",
            "type": "string"
          }
        },
        "required": [
          "sub_question",
          "sub_answer"
        ],
        "title": "SubQuestionAnswer",
        "type": "object"
      },
      "SubQuestionArtifacts": {
        "properties": {
          "citation_rows_by_index": {
            "additionalProperties": {
              "$ref": "#/components/schemas/CitationSourceRow"
            },
            "title": "Citation Rows By Index",
            "type": "object"
          },
          "expanded_queries": {
            "items": {
              "type": "string"
            },
            "title": "Expanded Queries",
            "type": "array"
          },
          "reranked_docs": {
            "items": {
              "$ref": "#/components/schemas/CitationSourceRow"
            },
            "title": "Reranked Docs",
            "type": "array"
          },
          "retrieval_provenance": {
            "items": {
              "type": "object"
            },
            "title": "Retrieval Provenance",
            "type": "array"
          },
          "retrieved_docs": {
            "items": {
              "$ref": "#/components/schemas/CitationSourceRow"
            },
            "title": "Retrieved Docs",
            "type": "array"
          },
          "sub_answer": {
            "default": "",
            "title": "Sub Answer",
            "type": "string"
          },
          "sub_question": {
            "title": "Sub Question",
            "type": "string"
          }
        },
        "required": [
          "sub_question"
        ],
        "title": "SubQuestionArtifacts",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "agent-search",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/agents/run": {
      "post": {
        "operationId": "runtime_agent_run_api_agents_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeAgentRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeAgentRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runtime Agent Run",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/run-async": {
      "post": {
        "operationId": "runtime_agent_run_async_api_agents_run_async_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeAgentRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeAgentRunAsyncStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runtime Agent Run Async",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/run-cancel/{job_id}": {
      "post": {
        "operationId": "runtime_agent_run_cancel_api_agents_run_cancel__job_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeAgentRunAsyncCancelResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runtime Agent Run Cancel",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/run-events/{job_id}": {
      "get": {
        "operationId": "runtime_agent_run_events_api_agents_run_events__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runtime Agent Run Events",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/run-resume/{job_id}": {
      "post": {
        "operationId": "runtime_agent_run_resume_api_agents_run_resume__job_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeAgentRunResumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeAgentRunAsyncStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runtime Agent Run Resume",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/agents/run-status/{job_id}": {
      "get": {
        "operationId": "runtime_agent_run_status_api_agents_run_status__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeAgentRunAsyncStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Runtime Agent Run Status",
        "tags": [
          "agents"
        ]
      }
    },
...
  }
}